requirements.txt
argparse
googletrans==4.0.0-rc1
ruamel.yaml
concurrent.futures